+++ packages/expo-splash-screen/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 🐛 Bug fixes
 
+- [Android] On Android 13 (API 33), skip `setOnExitAnimationListener` so dismissing the splash does not crash with `SurfaceControl.checkNotReleased()` when the app is backgrounded during the transition. ([#44243](https://github.com/expo/expo/pull/44243) by [@huextrat](https://github.com/huextrat))
+
 ### 💡 Others
 
 - Make `backgroundColor` plugin prop optional, defaulting to `#ffffff`. ([#44098](https://github.com/expo/expo/pull/44098) by [@zoontek](https://github.com/zoontek))